Split installing the build dependencies into its own step
Installing build dependencies can be a little bit time consuming
depending on the amount of dependencies to install. This is time during
which the user running the CLI has no information about what is going on.
With this change, the user will at least know what we are doing.

The issue about now knowing what is going on during the rpmbuild part
remains though.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>