No longer rely on RPC to start the container
Until now, we expected the RPC server to be running on the target
system. The client would then client to the RPC server, use it to pull
and start the builder container, then connect to the RPC server running
into that container to do the rest of the process (rebuild the srpm,
build the rpm)

With this commit, the client directly pull the container and start it on
the target system without going through the RPC server. So this commit
also removes the code to pull, start, clean the container via the RPC
server.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>